start converting Emitter to Kotlin...
